feat: create debug fn in debug.c

+++ lib/hash/includes/hashing.h
@@ -6,14 +6,13 @@
 /*   By: lravier <lravier@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/04/28 19:51:36 by lravier       #+#    #+#                 */
-/*   Updated: 2020/05/08 14:36:04 by kim           ########   odam.nl         */
+/*   Updated: 2020/05/08 15:41:48 by kim           ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef HASHING_H
 # define HASHING_H
 # define BASE_SIZE 50
-// # define BASE_SIZE 50
 # define PRIME_A 151
 # define PRIME_B 163
 // # define VAL_T t_room *
@@ -45,7 +44,7 @@ int					next_prime(int n);
 long long unsigned	ft_pow_ll(int base, int pow);
 unsigned			get_hash(char *key, t_table *ht, int attempt);
 unsigned			rehash(char *key, t_table *new_table);
-int					insert_ht(t_table *ht, char *key, void *value);
+ssize_t				insert_ht(t_table *ht, char *key, void *value);
 int					resize_up(t_table **ht);
 int					resize_down(t_table **ht);
 t_entry				*new_entry(char *key, void *value);
